notificaciones automaticas/debuggear
